Resolve / WontFix several TODOs
Show some kind of progress indicator here [while adding asset], for if
it's non-instant to add the asset? --- the "Saving..." overlay does
this.

Can we do this [verify that code and asset sync state is consistent]
without re-rendering whole thing when adding an asset? --- there is
now only one sync-state.

Some representation that the assets are syncing to storage [in Assets
info panel]? --- the "Saving..." overlay does this.

Add interactive elements to patch tables. --- done.

Move LoadingState somewhere central? --- have just one sync-state.

Eliminate dup'd code for loading-state? --- have just one sync-state.

Can we reduce flickering?  It's a bit distracting. --- everything is
loaded before we set sync-state to "Syncd".

Add 'id' / 'slug' to ITrackedTutorial for consistency checking ---
don't think there's scope for them to become inconsistent.

Is this [the IProjectSummary] the right place to note whether a project
is tracking a tutorial? --- seems to be working.

Work out when to revoke this URL [the ObjectURL created for
images]. --- note that it's done in clear().