Do not scroll-to-bottom on WebSocket info
It scrolls the whole tab-content DIV to the bottom, which messes up
the project asset tab.  I tried just making it scroll to bottom if the
editor web socket log was the active pane, but then tabbing back to
the project assets tab kept the scrolled position.  Not really worth
disrupting things for the minor convenience of seeing the bottom
websocket log entries.