</s>Improve focus behaviour
After performing (or cancelling) an operation from an asset card's
ccmenu, return focus to the asset card.  When launching the "upsert
hat block" modal for an update, give focus to the current event-kind.
Delay focus when deleting an asset; didn't get to the bottom of this,
but invoking the operation via keyboard ("Delete" key) and then
confirming did not leave focus correct under Cypress.